Fix warnings'

+++ src/huffman/huffman/io/binary-io.h
@@ -6,9 +6,9 @@
 
 namespace io {
 
-	u64 read_bits(unsigned nbits, io::InputStream& input);
+	u64 read_bits(u64 nbits, io::InputStream& input);
 
-	void write_bits(u64 value, unsigned nbits, io::OutputStream& output);
+	void write_bits(u64 value, u64 nbits, io::OutputStream& output);
 }
 
 #endif
